paczkomat_backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
dj-database-url
geopy
requests
python-dotenv
psycopg2-binary
uuid